Thêm logic - Thêm logic Npc cây đậu có 2 chức năng là thu hoạch đậu thần và nâng cấp cây đậu ( đã làm logic + đổi render theo từng cây đậu chuẩn bị cho logic thu hoạch đậu

+++ core/src/main/java/com/dang/dragonboy/du_lieu/DuLieuNguoiChoi.java
@@ -558,7 +558,7 @@ public class DuLieuNguoiChoi {
         this.GiamSatThuongNhanVat += PtGiamSatThuong;
         this.GiamSatThuongNhanVat = Math.min(this.GiamSatThuongNhanVat,85);
     }
-    public void tangVang(int soLuong) {
+    public void tangVang(long soLuong) {
         this.vang += soLuong;
     }
     public void tangNgoc(int soLuong) {
